refactor(api): 统一API路径前缀为/api/v1并优化相关结构
feat(controller): 添加系统健康检查接口和根路径响应
fix(interceptor): 修改JWT Token获取方式与FastAPI保持一致
style(model): 更新数据模型字段注释和类型定义
docs(config): 添加测试执行相关配置文档
chore(deps): 更新依赖项配置和数据库驱动

+++ automation-agent/platform-java-server/src/main/java/com/platform/platformserver/entity/ApiInfo.java
@@ -11,10 +11,10 @@ import lombok.EqualsAndHashCode;
 public class ApiInfo extends BaseEntity {
     
     @TableField("project_id")
-    private Integer projectId;
+    private Long projectId;
     
     @TableField("module_id")
-    private Integer moduleId;
+    private Long moduleId;
     
     @TableField("api_name")
     private String apiName;
